Create VKUploadBase
