+++ vksdk_library/src/main/java/com/vk/sdk/api/docs/VKUploadDocBase.java
@@ -21,25 +21,16 @@
 
 package com.vk.sdk.api.docs;
 
-import com.vk.sdk.VKSdk;
-import com.vk.sdk.api.VKError;
-import com.vk.sdk.api.VKRequest;
-import com.vk.sdk.api.VKResponse;
-import com.vk.sdk.api.httpClient.VKAbstractOperation;
+import com.vk.sdk.api.VKUploadBase;
 import com.vk.sdk.api.httpClient.VKHttpClient;
 import com.vk.sdk.api.httpClient.VKJsonOperation;
-import com.vk.sdk.api.httpClient.VKJsonOperation.VKJSONOperationCompleteListener;
-
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.File;
-import java.util.concurrent.ExecutorService;
 
 /**
  * Provides common part of document upload process
  */
-public abstract class VKUploadDocBase extends VKRequest {
+public abstract class VKUploadDocBase extends VKUploadBase {
     /**
      * ID of group to upload
      */
@@ -49,143 +40,15 @@ public abstract class VKUploadDocBase extends VKRequest {
      */
     protected File mDoc;
 
-    protected abstract VKRequest getServerRequest();
-
-    protected abstract VKRequest getSaveRequest(JSONObject response);
-
     /**
      * Creates a VKUploadDocBase empty instance.
      */
     public VKUploadDocBase() {
-        super(null);
+        super();
     }
 
     @Override
-    public VKAbstractOperation getOperation() {
-        return new VKUploadDocOperation();
+    protected VKJsonOperation getUploadOperation(String uploadUrl) {
+        return new VKJsonOperation(VKHttpClient.docUploadRequest(uploadUrl, mDoc));
     }
-
-    protected class VKUploadDocOperation extends VKAbstractOperation {
-		protected VKAbstractOperation lastOperation;
-
-        @Override
-        public void start(ExecutorService s) {
-            super.start(s);
-            final VKRequestListener originalListener = VKUploadDocBase.this.requestListener;
-
-            VKUploadDocBase.this.requestListener = new VKRequestListener() {
-                @Override
-                public void onComplete(VKResponse response) {
-                    setState(VKOperationState.Finished);
-                    response.request = VKUploadDocBase.this;
-                    if (originalListener != null) {
-                        originalListener.onComplete(response);
-                    }
-                }
-
-                @Override
-                public void onError(VKError error) {
-                    setState(VKOperationState.Finished);
-                    error.request = VKUploadDocBase.this;
-                    if (originalListener != null) {
-                        originalListener.onError(error);
-                    }
-                }
-
-                @Override
-                public void onProgress(VKProgressType progressType, long bytesLoaded,
-                                       long bytesTotal) {
-                    if (originalListener != null) {
-                        originalListener.onProgress(progressType, bytesLoaded, bytesTotal);
-                    }
-                }
-            };
-            setState(VKOperationState.Executing);
-
-            VKRequest serverRequest = getServerRequest();
-
-            serverRequest.setRequestListener(new VKRequestListener() {
-                @Override
-                public void onComplete(VKResponse response) {
-                    try {
-                        VKJsonOperation postFileRequest = new VKJsonOperation(
-                                VKHttpClient.docUploadRequest(response.json.getJSONObject("response").getString("upload_url"), mDoc));
-                        postFileRequest.setHttpOperationListener(new VKJSONOperationCompleteListener() {
-                            @Override
-                            public void onComplete(VKJsonOperation operation,
-                                                   JSONObject response) {
-
-                                VKRequest saveRequest = getSaveRequest(response);
-                                saveRequest.setRequestListener(new VKRequestListener() {
-                                    @Override
-                                    public void onComplete(VKResponse response) {
-                                        if (requestListener != null) {
-                                            requestListener.onComplete(response);
-                                        }
-                                        setState(VKOperationState.Finished);
-                                    }
-
-                                    @Override
-                                    public void onError(VKError error) {
-                                        if (requestListener != null) {
-                                            requestListener.onError(error);
-                                        }
-                                    }
-                                });
-                                lastOperation = saveRequest.getOperation();
-                                VKHttpClient.enqueueOperation(lastOperation);
-                            }
-
-                            @Override
-                            public void onError(VKJsonOperation operation, VKError error) {
-                                if (requestListener != null) {
-                                    requestListener.onError(error);
-                                }
-                            }
-                        });
-
-                        lastOperation = postFileRequest;
-                        VKHttpClient.enqueueOperation(lastOperation);
-                    } catch (JSONException e) {
-                        if (VKSdk.DEBUG)
-                            e.printStackTrace();
-                        VKError error = new VKError(VKError.VK_JSON_FAILED);
-                        error.httpError = e;
-                        error.errorMessage = e.getMessage();
-                        if (requestListener != null) {
-                            requestListener.onError(error);
-                        }
-                    }
-                }
-
-                @Override
-                public void onError(VKError error) {
-                    if (requestListener != null)
-                        requestListener.onError(error);
-                }
-            });
-            lastOperation = serverRequest.getOperation();
-            VKHttpClient.enqueueOperation(lastOperation);
-        }
-
-        @Override
-        public void cancel() {
-            if (lastOperation != null)
-                lastOperation.cancel();
-            super.cancel();
-        }
-
-        @Override
-        public void finish() {
-            super.finish();
-            lastOperation = null;
-        }
-
-        @Override
-        public Object getResultObject() {
-            return null;
-        }
-    }
-
-
 }
